+++ resources/lang/en/validation.php
@@ -90,7 +90,7 @@ return [
 
     'date_time_iso_8601' => 'The :attribute is not a valid ISO 8601 date and time.',
 
-    'disallowed_parameter' => 'Parameter :name is not allowed.',
+    'parameter_not_supported' => 'Parameter :name is not allowed.',
 
     'has_one' => 'The :attribute field must be a to-one relationship containing :types resources.',